Print the ticks as unsigned long long

+++ src/task.c
@@ -803,7 +803,8 @@ void task_dump_all(struct engine *e, int step) {
           file_thread, " %i %i %i %i %lli %lli %i %i %i %i %i\n",
           e->sched.tasks[l].rid, e->sched.tasks[l].type,
           e->sched.tasks[l].subtype, (e->sched.tasks[l].cj == NULL),
-          e->sched.tasks[l].tic, e->sched.tasks[l].toc,
+          (unsigned long long)e->sched.tasks[l].tic,
+          (unsigned long long)e->sched.tasks[l].toc,
           (e->sched.tasks[l].ci == NULL) ? 0
                                          : e->sched.tasks[l].ci->hydro.count,
           (e->sched.tasks[l].cj == NULL) ? 0
